Invoke TConn().init() in gunicorn worker thread.
MainThread in gunicorn gthread worker calls TConn().init() function,
but when ssl is used that function may raise an ENOTCONN exception.
There is very low probability for that, but still it is needed to catch
potential socket errors during that call, otherwise whole gunicorn worker
with all threads will be crashed.